feat(dashboards) Add editable default dashboards (#22472)
For pre-built dashboards I need to prepend ORM results with a static
list of results. In order to paginate this data in a somewhat sane
fashion we need a way iterate the static list and the dynamic one.

Include pre-built dashboards in the API responses for an organization's
dashboards. This will let us hide them via tombstones that are created
when pre-built dashboards are created or deleted.

+++ src/sentry/api/endpoints/organization_dashboard_details.py
@@ -6,6 +6,7 @@ from rest_framework.response import Response
 from sentry.api.bases.dashboard import OrganizationDashboardEndpoint
 from sentry.api.serializers import serialize
 from sentry.api.serializers.rest_framework import DashboardDetailsSerializer
+from sentry.models.dashboard import DashboardTombstone
 
 
 class OrganizationDashboardDetailsEndpoint(OrganizationDashboardEndpoint):
@@ -21,7 +22,8 @@ class OrganizationDashboardDetailsEndpoint(OrganizationDashboardEndpoint):
         :pparam int dashboard_id: the id of the dashboard.
         :auth: required
         """
-
+        if isinstance(dashboard, dict):
+            return self.respond(dashboard)
         return self.respond(serialize(dashboard, request.user))
 
     def delete(self, request, organization, dashboard):
@@ -29,14 +31,20 @@ class OrganizationDashboardDetailsEndpoint(OrganizationDashboardEndpoint):
         Delete an Organization's Dashboard
         ```````````````````````````````````
 
-        Delete an individual organization's dashboard.
+        Delete an individual organization's dashboard, or tombstone
+        a pre-built dashboard which effectively deletes it.
 
         :pparam string organization_slug: the slug of the organization the
                                           dashboard belongs to.
         :pparam int dashboard_id: the id of the dashboard.
         :auth: required
         """
-        dashboard.delete()
+        if isinstance(dashboard, dict):
+            DashboardTombstone.objects.get_or_create(
+                organization=organization, slug=dashboard["id"]
+            )
+        else:
+            dashboard.delete()
 
         return self.respond(status=204)
 
@@ -55,14 +63,27 @@ class OrganizationDashboardDetailsEndpoint(OrganizationDashboardEndpoint):
                             to be updated.
         :auth: required
         """
-        serializer = DashboardDetailsSerializer(data=request.data, instance=dashboard)
+        tombstone = None
+        if isinstance(dashboard, dict):
+            tombstone = dashboard["id"]
+            dashboard = None
+
+        serializer = DashboardDetailsSerializer(
+            data=request.data,
+            instance=dashboard,
+            context={"organization": organization, "request": request},
+        )
 
         if not serializer.is_valid():
             return Response(serializer.errors, status=400)
         try:
             with transaction.atomic():
                 serializer.save()
+                if tombstone:
+                    DashboardTombstone.objects.get_or_create(
+                        organization=organization, slug=tombstone
+                    )
         except IntegrityError:
             return self.respond({"Dashboard with that title already exists."}, status=409)
 
-        return self.respond(serialize(dashboard, request.user), status=200)
+        return self.respond(serialize(serializer.instance, request.user), status=200)
